requirements.txt
aiomysql==0.2.0
aiosqlite==0.20.0
discord.py==2.4.0
langdetect==1.0.9
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
translators==5.9.3